upload from linux

+++ q_2/q2_0/check_pid.c
@@ -2,12 +2,29 @@
 #include <stdio.h>
 #include <errno.h>
 #include <signal.h>
+#include <stdlib.h>
 
-int main()
+int isNumber(char* p ){
+    while (*p != '\0')
+    {
+        if (*p<'0' || *p>'9')
+        {
+            
+            return -1;
+        }
+        p++;
+    }
+    return 1;
+}
+
+int main(int agrc, char* argv[])
 {
-    int pid;
-    printf("check_pid ");
-    scanf("%d", &pid);
+
+    if(isNumber(argv[1])==-1){
+        printf("Invalid pid\n");
+        return 0;
+    }
+    int pid= atoi(argv[1]); //convert to int
     if (kill(pid,0)==-1)
        {
 	    if(errno==EPERM)
